eco_backend/requirements.txt
Flask==2.3.3
pydantic==2.6.4
mysql-connector-python==8.3.0
python-dotenv==1.0.1